feat(6742): extract topology mapping contract tests
