+++ crates/kamn-node/src/main_tests/daemon_tests/live_postgres_topology_contract_tests/topology_mapping_contract_tests.rs
@@ -1,768 +1,21 @@
-#[test]
-fn functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_pair_contract_is_canonical(
-) {
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_SCHEMA_VERSION,
-        "kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-pair.v1"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_REQUIRED_HOST_PAIR_IDS_CSV,
-        ["node_alpha->node_alpha", "node_alpha->node_beta"].join(",")
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_CONTRACT,
-        "host_pair_ids_must_remain_stable_under_repeated_runs_and_topology_permutations"
-    );
-
-    let sample_lane_fingerprint = format_parallel_lane_fingerprint(
-        "processor_listener_parallel_applied",
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-    );
-    let topology_fingerprint = format_parallel_lane_topology_fingerprint(
-        "same_host_parallel",
-        "node_alpha",
-        "node_alpha",
-        vec![sample_lane_fingerprint],
-    );
-    assert_eq!(
-        extract_parallel_lane_topology_host_pair_id(&topology_fingerprint),
-        "node_alpha->node_alpha"
-    );
-}
-
-#[test]
-fn integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_pairs_are_stable(
-) {
-    let _lock = log_env_lock()
-        .lock()
-        .unwrap_or_else(|error| error.into_inner());
-    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
-    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
-    let (gate_reason_code, maybe_database_url) = resolve_live_postgres_gate_decision();
-    let Some(database_url) = maybe_database_url else {
-        assert_eq!(gate_reason_code, LIVE_POSTGRES_ENV_UNSET_REASON_CODE);
-        return;
-    };
-    assert_eq!(
-        gate_reason_code,
-        LIVE_POSTGRES_ADAPTER_CONNECTED_REASON_CODE
-    );
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("tokio runtime should be constructible for live postgres validation");
-    runtime.block_on(async move {
-        let adapter = kamn_core::DataLayerPgExecutionAdapter::connect(
-            kamn_core::DataLayerPgExecutionAdapterConfig {
-                database_url,
-                max_connections: 4,
-            },
-        )
-        .await
-        .expect("live postgres connection should succeed when test URL is provided");
-        adapter
-            .apply_migrations()
-            .await
-            .expect("live postgres migrations should apply for validation slice");
-    });
-
-    let permutation_ids = ["baseline", "reverse", "rotate_left_1"];
-    let baseline_host_pair_ids =
-        collect_parallel_lane_topology_host_pair_ids(permute_parallel_lane_topology_profiles(
-            project_live_postgres_parallel_lane_topology_profiles(),
-            permutation_ids[0],
-        ));
-    assert_eq!(
-        baseline_host_pair_ids,
-        vec!["node_alpha->node_alpha", "node_alpha->node_beta"]
-    );
-    assert_eq!(
-        baseline_host_pair_ids.join(","),
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_REQUIRED_HOST_PAIR_IDS_CSV
-    );
-
-    for permutation in permutation_ids.iter().skip(1) {
-        let permuted_host_pair_ids =
-            collect_parallel_lane_topology_host_pair_ids(permute_parallel_lane_topology_profiles(
-                project_live_postgres_parallel_lane_topology_profiles(),
-                permutation,
-            ));
-        assert_eq!(
-            baseline_host_pair_ids, permuted_host_pair_ids,
-            "topology host-pair ids should remain stable under permutation {permutation}"
-        );
-    }
-}
-
-#[test]
-fn functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_pair_directionality_contract_is_canonical(
-) {
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_DIRECTIONALITY_SCHEMA_VERSION,
-        "kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-pair-directionality.v1"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_DIRECTIONALITY_EXTRACTION_RULE,
-        "host_a_to_host_b_arrow_notation_non_commutative"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_DIRECTIONALITY_FORBIDDEN_REVERSE_PAIRS_CSV,
-        "node_beta->node_alpha"
-    );
-
-    let sample_lane_fingerprint = format_parallel_lane_fingerprint(
-        "listener_approver_asymmetric_parallel_applied",
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-    );
-    let distributed_fingerprint = format_parallel_lane_topology_fingerprint(
-        "distributed_label_parallel",
-        "node_alpha",
-        "node_beta",
-        vec![sample_lane_fingerprint],
-    );
-    let canonical = extract_parallel_lane_topology_host_pair_id(&distributed_fingerprint);
-    let reverse = extract_parallel_lane_topology_host_pair_reverse_id(&distributed_fingerprint);
-    assert_eq!(canonical, "node_alpha->node_beta");
-    assert_eq!(reverse, "node_beta->node_alpha");
-    assert_ne!(
-        canonical, reverse,
-        "host-pair extraction should remain non-commutative"
-    );
-}
-
-#[test]
-fn integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_pair_directionality_is_stable(
-) {
-    let _lock = log_env_lock()
-        .lock()
-        .unwrap_or_else(|error| error.into_inner());
-    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
-    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
-    let (gate_reason_code, maybe_database_url) = resolve_live_postgres_gate_decision();
-    let Some(database_url) = maybe_database_url else {
-        assert_eq!(gate_reason_code, LIVE_POSTGRES_ENV_UNSET_REASON_CODE);
-        return;
-    };
-    assert_eq!(
-        gate_reason_code,
-        LIVE_POSTGRES_ADAPTER_CONNECTED_REASON_CODE
-    );
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("tokio runtime should be constructible for live postgres validation");
-    runtime.block_on(async move {
-        let adapter = kamn_core::DataLayerPgExecutionAdapter::connect(
-            kamn_core::DataLayerPgExecutionAdapterConfig {
-                database_url,
-                max_connections: 4,
-            },
-        )
-        .await
-        .expect("live postgres connection should succeed when test URL is provided");
-        adapter
-            .apply_migrations()
-            .await
-            .expect("live postgres migrations should apply for validation slice");
-    });
-
-    let permutation_ids = ["baseline", "reverse", "rotate_left_1"];
-    let baseline_topology_fingerprints =
-        run_parallel_lane_topology_fingerprints(permute_parallel_lane_topology_profiles(
-            project_live_postgres_parallel_lane_topology_profiles(),
-            permutation_ids[0],
-        ));
-    let mut baseline_canonical_ids = baseline_topology_fingerprints
-        .iter()
-        .map(|fingerprint| extract_parallel_lane_topology_host_pair_id(fingerprint))
-        .collect::<Vec<_>>();
-    baseline_canonical_ids.sort();
-    assert_eq!(
-        baseline_canonical_ids,
-        vec!["node_alpha->node_alpha", "node_alpha->node_beta"]
-    );
-
-    let forbidden_reverse_pairs =
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_DIRECTIONALITY_FORBIDDEN_REVERSE_PAIRS_CSV
-            .split(',')
-            .collect::<Vec<_>>();
-
-    for permutation in permutation_ids.iter().skip(1) {
-        let permuted_topology_fingerprints =
-            run_parallel_lane_topology_fingerprints(permute_parallel_lane_topology_profiles(
-                project_live_postgres_parallel_lane_topology_profiles(),
-                permutation,
-            ));
-        let mut permuted_canonical_ids = permuted_topology_fingerprints
-            .iter()
-            .map(|fingerprint| extract_parallel_lane_topology_host_pair_id(fingerprint))
-            .collect::<Vec<_>>();
-        permuted_canonical_ids.sort();
-        assert_eq!(
-            baseline_canonical_ids, permuted_canonical_ids,
-            "canonical host-pair ids should remain stable under permutation {permutation}"
-        );
-
-        for topology_fingerprint in &permuted_topology_fingerprints {
-            let reverse_id =
-                extract_parallel_lane_topology_host_pair_reverse_id(topology_fingerprint);
-            assert!(
-                !forbidden_reverse_pairs.contains(&reverse_id.as_str()),
-                "reverse host-pair id {} must remain forbidden under directionality contract",
-                reverse_id
-            );
-        }
-    }
-}
-
-#[test]
-fn functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_pair_mapping_contract_is_canonical(
-) {
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_MAPPING_SCHEMA_VERSION,
-        "kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-pair-mapping.v1"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_MAPPING_ROWS_CSV,
-        "same_host_parallel->node_alpha->node_alpha,distributed_label_parallel->node_alpha->node_beta"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_MAPPING_CONTRACT,
-        "topology_id_to_host_pair_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    );
-
-    let sample_lane_fingerprint = format_parallel_lane_fingerprint(
-        "processor_listener_parallel_applied",
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-    );
-    let topology_fingerprint = format_parallel_lane_topology_fingerprint(
-        "same_host_parallel",
-        "node_alpha",
-        "node_alpha",
-        vec![sample_lane_fingerprint],
-    );
-    assert_eq!(
-        extract_parallel_lane_topology_id_host_pair_row(&topology_fingerprint),
-        "same_host_parallel->node_alpha->node_alpha"
-    );
-}
-
-#[test]
-fn integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_pair_mapping_is_stable(
-) {
-    let _lock = log_env_lock()
-        .lock()
-        .unwrap_or_else(|error| error.into_inner());
-    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
-    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
-    let (gate_reason_code, maybe_database_url) = resolve_live_postgres_gate_decision();
-    let Some(database_url) = maybe_database_url else {
-        assert_eq!(gate_reason_code, LIVE_POSTGRES_ENV_UNSET_REASON_CODE);
-        return;
-    };
-    assert_eq!(
-        gate_reason_code,
-        LIVE_POSTGRES_ADAPTER_CONNECTED_REASON_CODE
-    );
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("tokio runtime should be constructible for live postgres validation");
-    runtime.block_on(async move {
-        let adapter = kamn_core::DataLayerPgExecutionAdapter::connect(
-            kamn_core::DataLayerPgExecutionAdapterConfig {
-                database_url,
-                max_connections: 4,
-            },
-        )
-        .await
-        .expect("live postgres connection should succeed when test URL is provided");
-        adapter
-            .apply_migrations()
-            .await
-            .expect("live postgres migrations should apply for validation slice");
-    });
-
-    let permutation_ids = ["baseline", "reverse", "rotate_left_1"];
-    let baseline_rows =
-        collect_parallel_lane_topology_id_host_pair_rows(permute_parallel_lane_topology_profiles(
-            project_live_postgres_parallel_lane_topology_profiles(),
-            permutation_ids[0],
-        ));
-    assert_eq!(
-        baseline_rows,
-        vec![
-            "distributed_label_parallel->node_alpha->node_beta",
-            "same_host_parallel->node_alpha->node_alpha"
-        ]
-    );
-    assert_eq!(
-        baseline_rows.join(","),
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_PAIR_MAPPING_ROWS_CSV
-    );
-
-    for permutation in permutation_ids.iter().skip(1) {
-        let permuted_rows = collect_parallel_lane_topology_id_host_pair_rows(
-            permute_parallel_lane_topology_profiles(
-                project_live_postgres_parallel_lane_topology_profiles(),
-                permutation,
-            ),
-        );
-        assert_eq!(
-            baseline_rows, permuted_rows,
-            "topology-id to host-pair rows should remain stable under permutation {permutation}"
-        );
-    }
-}
-
-#[test]
-fn functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_lane_set_mapping_contract_is_canonical(
-) {
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_SET_MAPPING_SCHEMA_VERSION,
-        "kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-lane-set-mapping.v1"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_SET_MAPPING_ROWS_CSV,
-        "same_host_parallel->symmetric_parallel,distributed_label_parallel->asymmetric_parallel"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_SET_MAPPING_CONTRACT,
-        "topology_id_to_lane_set_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    );
-
-    let sample_lane_fingerprint = format_parallel_lane_fingerprint(
-        "processor_listener_parallel_applied",
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-    );
-    let topology_fingerprint = format_parallel_lane_topology_fingerprint(
-        "same_host_parallel",
-        "node_alpha",
-        "node_alpha",
-        vec![sample_lane_fingerprint],
-    );
-    assert_eq!(
-        extract_parallel_lane_topology_id_lane_set_row(&topology_fingerprint),
-        "same_host_parallel->symmetric_parallel"
-    );
-}
-
-#[test]
-fn integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_lane_set_mapping_is_stable(
-) {
-    let _lock = log_env_lock()
-        .lock()
-        .unwrap_or_else(|error| error.into_inner());
-    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
-    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
-    let (gate_reason_code, maybe_database_url) = resolve_live_postgres_gate_decision();
-    let Some(database_url) = maybe_database_url else {
-        assert_eq!(gate_reason_code, LIVE_POSTGRES_ENV_UNSET_REASON_CODE);
-        return;
-    };
-    assert_eq!(
-        gate_reason_code,
-        LIVE_POSTGRES_ADAPTER_CONNECTED_REASON_CODE
-    );
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("tokio runtime should be constructible for live postgres validation");
-    runtime.block_on(async move {
-        let adapter = kamn_core::DataLayerPgExecutionAdapter::connect(
-            kamn_core::DataLayerPgExecutionAdapterConfig {
-                database_url,
-                max_connections: 4,
-            },
-        )
-        .await
-        .expect("live postgres connection should succeed when test URL is provided");
-        adapter
-            .apply_migrations()
-            .await
-            .expect("live postgres migrations should apply for validation slice");
-    });
-
-    let permutation_ids = ["baseline", "reverse", "rotate_left_1"];
-    let baseline_rows =
-        collect_parallel_lane_topology_id_lane_set_rows(permute_parallel_lane_topology_profiles(
-            project_live_postgres_parallel_lane_topology_profiles(),
-            permutation_ids[0],
-        ));
-    assert_eq!(
-        baseline_rows,
-        vec![
-            "distributed_label_parallel->asymmetric_parallel",
-            "same_host_parallel->symmetric_parallel"
-        ]
-    );
-    assert_eq!(
-        baseline_rows.join(","),
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_SET_MAPPING_ROWS_CSV
-    );
-
-    for permutation in permutation_ids.iter().skip(1) {
-        let permuted_rows = collect_parallel_lane_topology_id_lane_set_rows(
-            permute_parallel_lane_topology_profiles(
-                project_live_postgres_parallel_lane_topology_profiles(),
-                permutation,
-            ),
-        );
-        assert_eq!(
-            baseline_rows, permuted_rows,
-            "topology-id to lane-set rows should remain stable under permutation {permutation}"
-        );
-    }
-}
-
-#[test]
-fn functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_lane_count_mapping_contract_is_canonical(
-) {
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_COUNT_MAPPING_SCHEMA_VERSION,
-        "kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-lane-count-mapping.v1"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_COUNT_MAPPING_ROWS_CSV,
-        "same_host_parallel->4,distributed_label_parallel->4"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_COUNT_MAPPING_CONTRACT,
-        "topology_id_to_lane_count_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    );
-
-    let sample_lane_fingerprint = format_parallel_lane_fingerprint(
-        "processor_listener_parallel_applied",
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-    );
-    let topology_fingerprint = format_parallel_lane_topology_fingerprint(
-        "same_host_parallel",
-        "node_alpha",
-        "node_alpha",
-        vec![sample_lane_fingerprint],
-    );
-    assert_eq!(
-        extract_parallel_lane_topology_id_lane_count_row(&topology_fingerprint),
-        "same_host_parallel->1"
-    );
-}
-
-#[test]
-fn integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_lane_count_mapping_is_stable(
-) {
-    let _lock = log_env_lock()
-        .lock()
-        .unwrap_or_else(|error| error.into_inner());
-    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
-    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
-    let (gate_reason_code, maybe_database_url) = resolve_live_postgres_gate_decision();
-    let Some(database_url) = maybe_database_url else {
-        assert_eq!(gate_reason_code, LIVE_POSTGRES_ENV_UNSET_REASON_CODE);
-        return;
-    };
-    assert_eq!(
-        gate_reason_code,
-        LIVE_POSTGRES_ADAPTER_CONNECTED_REASON_CODE
-    );
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("tokio runtime should be constructible for live postgres validation");
-    runtime.block_on(async move {
-        let adapter = kamn_core::DataLayerPgExecutionAdapter::connect(
-            kamn_core::DataLayerPgExecutionAdapterConfig {
-                database_url,
-                max_connections: 4,
-            },
-        )
-        .await
-        .expect("live postgres connection should succeed when test URL is provided");
-        adapter
-            .apply_migrations()
-            .await
-            .expect("live postgres migrations should apply for validation slice");
-    });
-
-    let permutation_ids = ["baseline", "reverse", "rotate_left_1"];
-    let baseline_rows =
-        collect_parallel_lane_topology_id_lane_count_rows(permute_parallel_lane_topology_profiles(
-            project_live_postgres_parallel_lane_topology_profiles(),
-            permutation_ids[0],
-        ));
-    assert_eq!(
-        baseline_rows,
-        vec!["distributed_label_parallel->4", "same_host_parallel->4"]
-    );
-    assert_eq!(
-        baseline_rows.join(","),
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_LANE_COUNT_MAPPING_ROWS_CSV
-    );
-
-    for permutation in permutation_ids.iter().skip(1) {
-        let permuted_rows = collect_parallel_lane_topology_id_lane_count_rows(
-            permute_parallel_lane_topology_profiles(
-                project_live_postgres_parallel_lane_topology_profiles(),
-                permutation,
-            ),
-        );
-        assert_eq!(
-            baseline_rows, permuted_rows,
-            "topology-id to lane-count rows should remain stable under permutation {permutation}"
-        );
-    }
-}
-
-#[test]
-fn functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_mode_mapping_contract_is_canonical(
-) {
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_MODE_MAPPING_SCHEMA_VERSION,
-        "kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-mode-mapping.v1"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_MODE_MAPPING_ROWS_CSV,
-        "same_host_parallel->same_host,distributed_label_parallel->distributed_label"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_MODE_MAPPING_CONTRACT,
-        "topology_id_to_host_mode_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    );
-
-    let sample_lane_fingerprint = format_parallel_lane_fingerprint(
-        "processor_listener_parallel_applied",
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-    );
-    let topology_fingerprint = format_parallel_lane_topology_fingerprint(
-        "same_host_parallel",
-        "node_alpha",
-        "node_alpha",
-        vec![sample_lane_fingerprint],
-    );
-    assert_eq!(
-        extract_parallel_lane_topology_id_host_mode_row(&topology_fingerprint),
-        "same_host_parallel->same_host"
-    );
-}
-
-#[test]
-fn integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_mode_mapping_is_stable(
-) {
-    let _lock = log_env_lock()
-        .lock()
-        .unwrap_or_else(|error| error.into_inner());
-    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
-    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
-    let (gate_reason_code, maybe_database_url) = resolve_live_postgres_gate_decision();
-    let Some(database_url) = maybe_database_url else {
-        assert_eq!(gate_reason_code, LIVE_POSTGRES_ENV_UNSET_REASON_CODE);
-        return;
-    };
-    assert_eq!(
-        gate_reason_code,
-        LIVE_POSTGRES_ADAPTER_CONNECTED_REASON_CODE
-    );
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("tokio runtime should be constructible for live postgres validation");
-    runtime.block_on(async move {
-        let adapter = kamn_core::DataLayerPgExecutionAdapter::connect(
-            kamn_core::DataLayerPgExecutionAdapterConfig {
-                database_url,
-                max_connections: 4,
-            },
-        )
-        .await
-        .expect("live postgres connection should succeed when test URL is provided");
-        adapter
-            .apply_migrations()
-            .await
-            .expect("live postgres migrations should apply for validation slice");
-    });
-
-    let permutation_ids = ["baseline", "reverse", "rotate_left_1"];
-    let baseline_rows =
-        collect_parallel_lane_topology_id_host_mode_rows(permute_parallel_lane_topology_profiles(
-            project_live_postgres_parallel_lane_topology_profiles(),
-            permutation_ids[0],
-        ));
-    assert_eq!(
-        baseline_rows,
-        vec![
-            "distributed_label_parallel->distributed_label",
-            "same_host_parallel->same_host"
-        ]
-    );
-    assert_eq!(
-        baseline_rows.join(","),
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_MODE_MAPPING_ROWS_CSV
-    );
-
-    for permutation in permutation_ids.iter().skip(1) {
-        let permuted_rows = collect_parallel_lane_topology_id_host_mode_rows(
-            permute_parallel_lane_topology_profiles(
-                project_live_postgres_parallel_lane_topology_profiles(),
-                permutation,
-            ),
-        );
-        assert_eq!(
-            baseline_rows, permuted_rows,
-            "topology-id to host-mode rows should remain stable under permutation {permutation}"
-        );
-    }
-}
-
-#[test]
-fn functional_runtime_daemon_live_postgres_validation_slice_parallel_lane_topology_host_cardinality_mapping_contract_is_canonical(
-) {
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_CARDINALITY_MAPPING_SCHEMA_VERSION,
-        "kamn.runtime.daemon.phase6-live-postgres.parallel-lane-topology-host-cardinality-mapping.v1"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_CARDINALITY_MAPPING_ROWS_CSV,
-        "same_host_parallel->1,distributed_label_parallel->2"
-    );
-    assert_eq!(
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_CARDINALITY_MAPPING_CONTRACT,
-        "topology_id_to_unique_host_cardinality_rows_must_remain_stable_under_repeated_runs_and_permutations"
-    );
-
-    let sample_lane_fingerprint = format_parallel_lane_fingerprint(
-        "processor_listener_parallel_applied",
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-        &LivePostgresPhase6Projection {
-            reason_code: LIVE_POSTGRES_MATRIX_PHASE6_APPLIED_REASON_CODE.to_owned(),
-            reason_taxonomy_version: LIVE_POSTGRES_DAEMON_REASON_TAXONOMY_VERSION.to_owned(),
-        },
-    );
-    let topology_fingerprint = format_parallel_lane_topology_fingerprint(
-        "distributed_label_parallel",
-        "node_alpha",
-        "node_beta",
-        vec![sample_lane_fingerprint],
-    );
-    assert_eq!(
-        extract_parallel_lane_topology_id_host_cardinality_row(&topology_fingerprint),
-        "distributed_label_parallel->2"
-    );
-}
-
-#[test]
-fn integration_runtime_daemon_phase6_live_postgres_validation_slice_parallel_lane_topology_host_cardinality_mapping_is_stable(
-) {
-    let _lock = log_env_lock()
-        .lock()
-        .unwrap_or_else(|error| error.into_inner());
-    let _level_guard = EnvVarGuard::set("KAMN_NODE_LOG_LEVEL", Some("info"));
-    let _format_guard = EnvVarGuard::set("KAMN_NODE_LOG_FORMAT", Some("json"));
-    let (gate_reason_code, maybe_database_url) = resolve_live_postgres_gate_decision();
-    let Some(database_url) = maybe_database_url else {
-        assert_eq!(gate_reason_code, LIVE_POSTGRES_ENV_UNSET_REASON_CODE);
-        return;
-    };
-    assert_eq!(
-        gate_reason_code,
-        LIVE_POSTGRES_ADAPTER_CONNECTED_REASON_CODE
-    );
-
-    let runtime = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .expect("tokio runtime should be constructible for live postgres validation");
-    runtime.block_on(async move {
-        let adapter = kamn_core::DataLayerPgExecutionAdapter::connect(
-            kamn_core::DataLayerPgExecutionAdapterConfig {
-                database_url,
-                max_connections: 4,
-            },
-        )
-        .await
-        .expect("live postgres connection should succeed when test URL is provided");
-        adapter
-            .apply_migrations()
-            .await
-            .expect("live postgres migrations should apply for validation slice");
-    });
-
-    let permutation_ids = ["baseline", "reverse", "rotate_left_1"];
-    let baseline_rows = collect_parallel_lane_topology_id_host_cardinality_rows(
-        permute_parallel_lane_topology_profiles(
-            project_live_postgres_parallel_lane_topology_profiles(),
-            permutation_ids[0],
-        ),
-    );
-    assert_eq!(
-        baseline_rows,
-        vec!["distributed_label_parallel->2", "same_host_parallel->1"]
-    );
-    assert_eq!(
-        baseline_rows.join(","),
-        LIVE_POSTGRES_PARALLEL_LANE_TOPOLOGY_HOST_CARDINALITY_MAPPING_ROWS_CSV
-    );
-
-    for permutation in permutation_ids.iter().skip(1) {
-        let permuted_rows = collect_parallel_lane_topology_id_host_cardinality_rows(
-            permute_parallel_lane_topology_profiles(
-                project_live_postgres_parallel_lane_topology_profiles(),
-                permutation,
-            ),
-        );
-        assert_eq!(
-            baseline_rows, permuted_rows,
-            "topology-id to host-cardinality rows should remain stable under permutation {permutation}"
-        );
-    }
-}
-
+// include!("live_postgres_topology_contract_tests/topology_mapping_contract_tests/host_pair_identity_contract_tests.rs");
+#[path = "topology_mapping_contract_tests/host_pair_identity_contract_tests.rs"]
+mod host_pair_identity_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_mapping_contract_tests/host_pair_directionality_contract_tests.rs");
+#[path = "topology_mapping_contract_tests/host_pair_directionality_contract_tests.rs"]
+mod host_pair_directionality_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_mapping_contract_tests/host_pair_mapping_contract_tests.rs");
+#[path = "topology_mapping_contract_tests/host_pair_mapping_contract_tests.rs"]
+mod host_pair_mapping_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_mapping_contract_tests/lane_set_mapping_contract_tests.rs");
+#[path = "topology_mapping_contract_tests/lane_set_mapping_contract_tests.rs"]
+mod lane_set_mapping_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_mapping_contract_tests/lane_count_mapping_contract_tests.rs");
+#[path = "topology_mapping_contract_tests/lane_count_mapping_contract_tests.rs"]
+mod lane_count_mapping_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_mapping_contract_tests/host_mode_mapping_contract_tests.rs");
+#[path = "topology_mapping_contract_tests/host_mode_mapping_contract_tests.rs"]
+mod host_mode_mapping_contract_tests;
+// include!("live_postgres_topology_contract_tests/topology_mapping_contract_tests/host_cardinality_mapping_contract_tests.rs");
+#[path = "topology_mapping_contract_tests/host_cardinality_mapping_contract_tests.rs"]
+mod host_cardinality_mapping_contract_tests;
